Further optimization of attack patterns
