+++ BasicMechanicsDemo/Assets/Scripts/Character/Mobile/Enemy/AttackPattern/MeleeAttackPattern.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: dbbfcb69028cd44668cd871779da2baa
+timeCreated: 1510916081
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
