+++ BasicMechanicsDemo/Assets/Scripts/Character/Mobile/Enemy/AttackPattern/RangedAttackPattern.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: f7a46d346bb684dea950710882ce427b
+timeCreated: 1510916095
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
